Fix float price

+++ web/controllers/c_buy_place.php
@@ -20,7 +20,7 @@ $line = intval($_POST['line']);
 $from = intval($_POST['from']);
 $to = intval($_POST['to']);
 $nbr = intval($_POST['nbr']);
-$price = intval($_POST['price']);
+$price = floatval($_POST['price']);
 $date = $_POST['date'];
 
 $exist = false;
